Work around `find_crate_name` issue in examples
For the examples in the `examples` directory, they access the main crate
using `rustorm`, but since they share the manifest with the main crate,
`find_crate_name` ends up thinking that the way to access `rustorm` is
using `crate`.

To work around this limitation, the imports from `rustorm` required by
the generated code need to be imported into the root of the example name
space.